+++ src/App.tsx
@@ -8,6 +8,7 @@ import { Paper } from './pages/Paper'
 import { Protocols } from './pages/Protocols'
 import { References } from './pages/References'
 import { ReviewPacket } from './pages/ReviewPacket'
+import { TrustNetwork } from './pages/TrustNetwork'
 import { ToyPage } from './pages/ToyPage'
 
 export default function App() {
@@ -17,6 +18,7 @@ export default function App() {
         <Route index element={<Overview />} />
         <Route path="paper" element={<Paper />} />
         <Route path="toy" element={<ToyPage />} />
+        <Route path="trust-network" element={<TrustNetwork />} />
         <Route path="protocols" element={<Protocols />} />
         <Route path="methods" element={<Methods />} />
         <Route path="faq" element={<FAQ />} />